rename(examples/MC_Truth): rename source and executable to GPUMCTruth
Match the naming convention of sibling binaries (GPUCerenkov,
GPURaytrace, GPUPhotonSource, ...). The new name also mirrors the
OPTICKS_MC_TRUTH env var that gates the feature, so source, build
artifact, and runtime toggle share vocabulary.

Updates:
* Source files: GPURaytrace.{cpp,h} -> GPUMCTruth.{cpp,h}
* CMakeLists.txt: project + add_executable + install target
* GPUMCTruth.cpp: usage string + include path
* run.sh: BIN path + header comment

Directory name examples/MC_Truth/ unchanged for now.
