+++ examples/MC_Truth/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.20)
 
-project(MC_Truth_GPURaytrace LANGUAGES CXX)
+project(GPUMCTruth LANGUAGES CXX)
 
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
@@ -10,12 +10,12 @@ find_package(glm REQUIRED)
 find_package(eic-opticks REQUIRED)
 find_package(Geant4 REQUIRED ui_all vis_all)
 
-add_executable(MC_Truth_GPURaytrace GPURaytrace.cpp GPURaytrace.h)
-target_link_libraries(MC_Truth_GPURaytrace
+add_executable(GPUMCTruth GPUMCTruth.cpp GPUMCTruth.h)
+target_link_libraries(GPUMCTruth
     eic-opticks::G4CX
     eic-opticks::SysRap
     eic-opticks::U4
     ${Geant4_LIBRARIES}
 )
 
-install(TARGETS MC_Truth_GPURaytrace)
+install(TARGETS GPUMCTruth)
